+++ app/services/wokbaskets/measure_association_savers/footnotes.rb
