+++ app/value_objects/workbaskets/shared/candidates_validations_summarizer.rb
