+++ app/value_objects/workbaskets/shared/system_ops_assigner.rb
@@ -1,5 +1,5 @@
-module CreateMeasures
-  module ValidationHelpers
+module Workbaskets
+  module Shared
     class SystemOpsAssigner
 
       attr_accessor :record,
